requirements.txt
psycopg[binary]>=3.1
pandas>=1.3
python-dotenv>=1.0